Move impact actor type and validity against terrain to warhead.
